modif DELETE agence/assurance/client/destination

+++ Projet Web/Controllers/AssurancesController.cs
@@ -96,6 +96,8 @@ namespace Projet_Web.Controllers
             {
                 return NotFound();
             }
+            if (assurance.DossierReservations == null)
+                return BadRequest("à un dossier");
 
             db.Assurances.Remove(assurance);
             db.SaveChanges();
